# src/main/resources/openapi/api-docs.yml
openapi: 3.0.0
info:
  title: Reviev platform - OpenAPI 1.0
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/platform
...
paths:
...
  /api/v1/company:
    post:
      operationId: createCompany
      summary: Create new company
      tags:
        - Company
      parameters:
        - in: header
          name: authorization
          required: true
          schema:
            type: string
      requestBody:
        description: Company DTO
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShortCompanyDto'
        required: true
      responses:
        '201':
          description: Created company
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyDto'
        '400':
          description: Invalid request
        '404':
          description: Resource not found
        '405':
          description: Method is not available
        '500':
          description: Server error
  /api/v1/companies/page={page}&size={size}:
    get:
      tags:
        - Company
      summary: Find all company
      operationId: findAllCompanys
      parameters:
        - in: header
          name: authorization
          required: true
          schema:
            type: string
        - name: page
          in: path
          required: true
          schema:
            type: integer
            nullable: false
        - name: size
          in: path
          required: true
          schema:
            type: integer
            nullable: false
      responses:
        '200':
          description: Get all company
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyListDto'
        '400':
          description: Invalid request
        '404':
          description: Resource not found
        '405':
          description: Method is not available
        '500':
          description: Server error
  /api/v1/company/{companyId}:
    get:
      tags:
        - Company
      summary: Find company by id
      operationId: findCompanyById
      parameters:
        - in: header
          name: authorization
          required: true
          schema:
            type: string
        - name: companyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            nullable: false
      responses:
        '200':
          description: Get company
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyDto'
        '400':
          description: Invalid request
        '404':
          description: Resource not found
        '405':
          description: Method is not available
        '500':
          description: Server error
    delete:
      tags:
        - Company
      summary: Delete company
      operationId: deleteCompany
      parameters:
        - in: header
          name: authorization
          required: true
          schema:
            type: string
        - name: companyId
          in: path
          description: Company id to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Delete company
        '400':
          description: Invalid request
        '404':
          description: Resource not found
        '405':
          description: Method is not available
        '500':
          description: Server error
    put:
      operationId: updateCompany
      summary: Update company
      tags:
        - Company
      parameters:
        - in: header
          name: authorization
          required: true
          schema:
            type: string
        - name: companyId
          in: path
          description: Company id to update
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: Company DTO
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyDto'
        required: true
      responses:
        '200':
          description: Updated company
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyDto'
        '400':
          description: Invalid request
        '404':
          description: Resource not found
        '405':
          description: Method is not available
        '500':
          description: Server error
  /api/v1/interview:
    post:
      operationId: createInterview
      summary: Create new interview
      tags:
        - Interview
      parameters:
        - in: header
          name: authorization
          required: true
          schema:
            type: string
      requestBody:
        description: Interview DTO
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShortInterviewDto'
        required: true
      responses:
        '201':
          description: Created interview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterviewDto'
        '400':
          description: Invalid request
        '404':
          description: Resource not found
        '405':
          description: Method is not available
        '500':
          description: Server error
  /api/v1/interviews/page={page}&size={size}:
    get:
      tags:
        - Interview
      summary: Find all interview
      operationId: findAllInterviews
      parameters:
        - in: header
          name: authorization
          required: true
          schema:
            type: string
        - name: page
          in: path
          required: true
          schema:
            type: integer
            nullable: false
        - name: size
          in: path
          required: true
          schema:
            type: integer
            nullable: false
      responses:
        '200':
          description: Get all interview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterviewListDto'
        '400':
          description: Invalid request
        '404':
          description: Resource not found
        '405':
          description: Method is not available
        '500':
          description: Server error
  /api/v1/interview/{interviewId}:
    get:
      tags:
        - Interview
      summary: Find interview by id
      operationId: findInterviewById
      parameters:
        - in: header
          name: authorization
          required: true
          schema:
            type: string
        - name: interviewId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            nullable: false
      responses:
        '200':
          description: Get interview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterviewDto'
        '400':
          description: Invalid request
        '404':
          description: Resource not found
        '405':
          description: Method is not available
        '500':
          description: Server error
    delete:
      tags:
        - Interview
      summary: Delete interview
      operationId: deleteInterview
      parameters:
        - in: header
          name: authorization
          required: true
          schema:
            type: string
        - name: interviewId
          in: path
          description: Company id to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Delete interview
        '400':
          description: Invalid request
        '404':
          description: Resource not found
        '405':
          description: Method is not available
        '500':
          description: Server error
    put:
      operationId: updateInterview
      summary: Update interview
      tags:
        - Interview
      parameters:
        - in: header
          name: authorization
          required: true
          schema:
            type: string
        - name: interviewId
          in: path
          description: Interview by id to update
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: Interview DTO
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InterviewDto'
        required: true
      responses:
        '200':
          description: Interview updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterviewDto'
        '400':
          description: Invalid request
        '404':
          description: Resource not found
        '405':
          description: Method is not available
        '500':
          description: Server error
...
  /api/v1/comment:
    post:
      operationId: createCommentary
      summary: Create new Commentary
      tags:
        - Commentary
      parameters:
        - in: header
          name: authorization
          required: true
          schema:
            type: string
      requestBody:
        description: Commentary DTO
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShortCommentaryDto'
        required: true
      responses:
        '201':
          description: Created Commentary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentaryDto'
        '400':
          description: Invalid request
        '404':
          description: Resource not found
        '405':
          description: Method is not available
        '500':
          description: Server error

    put:
      operationId: updateCommentary
      summary: Update commentary
      tags:
        - Commentary
      parameters:
        - in: header
          name: authorization
          required: true
          schema:
            type: string
      requestBody:
        description: Review DTO
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditCommentaryDto'
        required: true
      responses:
        '200':
          description: Review updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentaryDto'
        '400':
          description: Invalid request
        '404':
          description: Resource not found
        '405':
          description: Method is not available
        '500':
          description: Server error
  /api/v1/comment/{commentId}:
    delete:
      operationId: deleteCommentary
      summary: Delete commentary
      tags:
        - Commentary
      parameters:
        - in: header
          name: authorization
          required: true
          schema:
            type: string
        - name: commentId
          in: path
          description: Commentary id to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Delete commentary
        '400':
          description: Invalid request
        '404':
          description: Resource not found
        '405':
          description: Method is not available
        '500':
          description: Server error
  /api/v1/comment/admin/{commentId}:
    delete:
      operationId: deleteCommentaryFromDb
      summary: Delete commentary drom DB
      tags:
        - Commentary
      parameters:
        - in: header
          name: authorization
          required: true
          schema:
            type: string
        - name: commentId
          in: path
          description: Commentary id to delete from Db
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Delete commentary
        '400':
          description: Invalid request
        '404':
          description: Resource not found
        '405':
          description: Method is not available
        '500':
          description: Server error

# --- Dtos ---
components:
  schemas:
...
    UserDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: User id
        firstName:
          type: string
          description: User first name
        lastName:
          type: string
          description: User last name
        avatar:
          type: string
          description: User avatar url
        email:
          type: string
          description: User email
        role:
          type: string
          enum:
            - USER
            - ADMIN
        specialities:
          type: string
          enum:
            - FRONTEND
            - BACKEND
        reviewsReceiving:
          type: array
          items:
            $ref: '#/components/schemas/ReviewDto'
        reviewsStudent:
          type: array
          items:
            $ref: '#/components/schemas/ReviewDto'
        socials:
          type: array
          items:
            $ref: '#/components/schemas/SocialDto'
...
    SocialDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Social id
        name:
          type: string
          description: Social name
        link:
          type: string
          description: Social link
    CommentaryDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Commentary id
        user:
          $ref: '#/components/schemas/UserDto'
        text:
          type: string
        likes:
          type: number
        dislikes:
          type: number
        answers:
          type: array
          items:
            $ref: '#/components/schemas/CommentaryDto'
        create:
          type: string
          format: date-time
    ShortCommentaryDto:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        companyId:
          type: string
          format: uuid
        interviewId:
          type: string
          format: uuid
        subCommentaryId:
          type: string
          format: uuid
        text:
          type: string
    EditCommentaryDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        text:
          type: string
...
    CompanyDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Company id
        name:
          type: string
          description: Company name
        jobLink:
          type: string
          description: Company jobLink
        rating:
          type: number
          description: Company rating
        commentaries:
          type: array
          items:
            $ref: '#/components/schemas/CommentaryDto'
          description: Company commentaries
        interviews:
          type: array
          items:
            $ref: '#/components/schemas/InterviewDto'
          description: Company interviews
    ShortCompanyDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Social id
        name:
          type: string
          description: Social name
        link:
          type: string
          description: Social link
...
    CompanyListDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/CompanyDto'
        totalElements:
          type: number
        totalPages:
          type: number
        currentPage:
          type: number
        last:
          type: boolean
        numberOfElements:
          type: number
        first:
          type: boolean
        empty:
          type: boolean

    #   --Interview DTO--
    InterviewDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Interview id
        jobTitle:
          type: string
          description: Interview jobTitle
        jobLink:
          type: string
          description: Interview jobLink
        description:
          type: string
          description: Interview description
        money:
          type: integer
          description: Interview money
        videoLink:
          type: string
          description: Interview videoLink
        structure:
          type: string
          description: Interview structure
        substructure:
          type: string
          description: Interview substructure
        date:
          type: string
          format: date-time
          description: Interview dateInterview
        company:
          $ref: '#/components/schemas/CompanyDto'
        user:
          $ref: '#/components/schemas/UserDto'
        commentaries:
          type: array
          items:
            $ref: '#/components/schemas/CommentaryDto'
    ShortInterviewDto:
      type: object
      properties:
        jobTitle:
          type: string
          description: Interview jobTitle
        jobLink:
          type: string
          description: Interview jobLink
        description:
          type: string
          description: Interview description
        money:
          type: integer
          description: Interview money
        videoLink:
          type: string
          description: Interview videoLink
        structure:
          type: string
          description: Interview structure
        substructure:
          type: string
          description: Interview substructure
        date:
          type: string
          format: date-time
          description: Interview dateInterview
        company:
          $ref: '#/components/schemas/CompanyDto'
        user:
          $ref: '#/components/schemas/UserDto'
        authData:
          $ref: '#/components/schemas/AuthenticationToken'
    InterviewListDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/InterviewDto'
        totalElements:
          type: number
        totalPages:
          type: number
        currentPage:
          type: number
        last:
          type: boolean
        numberOfElements:
          type: number
        first:
          type: boolean
        empty:
          type: boolean

    #   --ReviewDto DTO--
    ReviewDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        theme:
          type: string
        time:
          type: string
          format: date-time
        link:
          type: string
        speciality:
          type: string
          enum:
            - FRONTEND
            - BACKEND
        status:
          type: string
          enum:
            - TOBE
            - PASSED
            - CANCELED
        reviewer:
          $ref: '#/components/schemas/UserDto'
        student:
          $ref: '#/components/schemas/UserDto'
...
    AuthenticationToken:
      type: object
      properties:
        token:
          type: string
          description: JWT